requirements.txt
streamlit
python-dotenv
openai
requests
google-generativeai
ollama
